requirements.txt
langchain==0.1.11
langchain_google_genai==0.0.9
python-dotenv==1.0.1
streamlit ==1.27.0
